Adding sequencer module to uber-program
Not complete and rather different that the current sequencer.
For instance it generates a waveform (sine) and not data for a DCO module.
The edition is missing parameter sync for the frequency.

The uber-program starts suffering from low memory mainly because of several "large" buffers. We will probably have to share them and store some values on program memory.